Applying review comments
